+++ src/pczt.rs
@@ -1,0 +1,297 @@
+//! PCZT support for Sapling.
+
+use core::fmt;
+use std::collections::BTreeMap;
+
+use getset::Getters;
+use redjubjub::{Binding, SpendAuth};
+use zcash_note_encryption::{
+    EphemeralKeyBytes, OutgoingCipherKey, ENC_CIPHERTEXT_SIZE, OUT_CIPHERTEXT_SIZE,
+};
+use zip32::ChildIndex;
+
+use crate::{
+    bundle::GrothProofBytes,
+    keys::SpendAuthorizingKey,
+    note::ExtractedNoteCommitment,
+    value::{NoteValue, ValueCommitTrapdoor, ValueCommitment, ValueSum},
+    Anchor, MerklePath, Nullifier, PaymentAddress, ProofGenerationKey, Rseed,
+};
+
+mod parse;
+pub use parse::ParseError;
+
+mod io_finalizer;
+pub use io_finalizer::IoFinalizerError;
+
+mod updater;
+pub use updater::{OutputUpdater, SpendUpdater, Updater, UpdaterError};
+
+mod prover;
+pub use prover::ProverError;
+
+mod signer;
+pub use signer::SignerError;
+
+mod tx_extractor;
+pub use tx_extractor::{TxExtractorError, Unbound};
+
+/// PCZT fields that are specific to producing the transaction's Sapling bundle (if any).
+///
+/// This struct is for representing Sapling in a partially-created transaction. If you
+/// have a fully-created transaction, use [the regular `Bundle` struct].
+///
+/// [the regular `Bundle` struct]: crate::Bundle
+#[derive(Debug, Getters)]
+#[getset(get = "pub")]
+pub struct Bundle {
+    /// The Sapling spends in this bundle.
+    ///
+    /// Entries are added by the Constructor, and modified by an Updater, IO Finalizer,
+    /// Prover, Signer, Combiner, or Spend Finalizer.
+    pub(crate) spends: Vec<Spend>,
+
+    /// The Sapling outputs in this bundle.
+    ///
+    /// Entries are added by the Constructor, and modified by an Updater, IO Finalizer,
+    /// Prover, Signer, Combiner, or Spend Finalizer.
+    pub(crate) outputs: Vec<Output>,
+
+    /// The net value of Sapling `spends` minus `outputs`.
+    ///
+    /// This is initialized by the Creator, and updated by the Constructor as spends or
+    /// outputs are added to the PCZT. It enables per-spend and per-output values to be
+    /// redacted from the PCZT after they are no longer necessary.
+    pub(crate) value_sum: ValueSum,
+
+    /// The Sapling anchor for this transaction.
+    ///
+    /// Set by the Creator.
+    pub(crate) anchor: Anchor,
+
+    /// The Sapling binding signature signing key.
+    ///
+    /// - This is `None` until it is set by the IO Finalizer.
+    /// - The Transaction Extractor uses this to produce the binding signature.
+    pub(crate) bsk: Option<redjubjub::SigningKey<Binding>>,
+}
+
+impl Bundle {
+    /// Returns a mutable reference to the spends in this bundle.
+    ///
+    /// This is used by Signers to apply signatures with [`Spend::sign`].
+    pub fn spends_mut(&mut self) -> &mut [Spend] {
+        &mut self.spends
+    }
+}
+
+/// Information about a Sapling spend within a transaction.
+///
+/// This struct is for representing Sapling spends in a partially-created transaction. If
+/// you have a fully-created transaction, use [the regular `SpendDescription` struct].
+///
+/// [the regular `SpendDescription` struct]: crate::bundle::SpendDescription
+#[derive(Debug, Getters)]
+#[getset(get = "pub")]
+pub struct Spend {
+    /// A commitment to the value consumed by this spend.
+    pub(crate) cv: ValueCommitment,
+
+    /// The nullifier of the note being spent.
+    pub(crate) nullifier: Nullifier,
+
+    /// The randomized verification key for the note being spent.
+    pub(crate) rk: redjubjub::VerificationKey<SpendAuth>,
+
+    /// The Spend proof.
+    ///
+    /// This is set by the Prover.
+    pub(crate) zkproof: Option<GrothProofBytes>,
+
+    /// The spend authorization signature.
+    ///
+    /// This is set by the Signer.
+    pub(crate) spend_auth_sig: Option<redjubjub::Signature<SpendAuth>>,
+
+    /// The address that received the note being spent.
+    ///
+    /// - This is set by the Constructor (or Updater?).
+    /// - This is required by the Prover.
+    pub(crate) recipient: Option<PaymentAddress>,
+
+    /// The value of the input being spent.
+    ///
+    /// This may be used by Signers to verify that the value matches `cv`, and to confirm
+    /// the values and change involved in the transaction.
+    ///
+    /// This exposes the input value to all participants. For Signers who don't need this
+    /// information, or after signatures have been applied, this can be redacted.
+    pub(crate) value: Option<NoteValue>,
+
+    /// The seed randomness for the note being spent.
+    ///
+    /// - This is set by the Constructor.
+    /// - This is required by the Prover.
+    pub(crate) rseed: Option<Rseed>,
+
+    /// The value commitment randomness.
+    ///
+    /// - This is set by the Constructor.
+    /// - The IO Finalizer compresses it into the `bsk`.
+    /// - This is required by the Prover.
+    /// - This may be used by Signers to verify that the value correctly matches `cv`.
+    ///
+    /// This opens `cv` for all participants. For Signers who don't need this information,
+    /// or after proofs / signatures have been applied, this can be redacted.
+    pub(crate) rcv: Option<ValueCommitTrapdoor>,
+
+    /// The proof generation key `(ak, nsk)` corresponding to the recipient that received
+    /// the note being spent.
+    ///
+    /// - This is set by the Updater.
+    /// - This is required by the Prover.
+    pub(crate) proof_generation_key: Option<ProofGenerationKey>,
+
+    /// A witness from the note to the bundle's anchor.
+    ///
+    /// - This is set by the Updater.
+    /// - This is required by the Prover.
+    pub(crate) witness: Option<MerklePath>,
+
+    /// The spend authorization randomizer.
+    ///
+    /// - This is chosen by the Constructor.
+    /// - This is required by the Signer for creating `spend_auth_sig`, and may be used to
+    ///   validate `rk`.
+    /// - After`zkproof` / `spend_auth_sig` has been set, this can be redacted.
+    pub(crate) alpha: Option<jubjub::Scalar>,
+
+    /// The ZIP 32 derivation path at which the spending key can be found for the note
+    /// being spent.
+    pub(crate) zip32_derivation: Option<Zip32Derivation>,
+
+    /// The spend authorizing key for this spent note, if it is a dummy note.
+    ///
+    /// - This is chosen by the Constructor.
+    /// - This is required by the IO Finalizer, and is cleared by it once used.
+    /// - Signers MUST reject PCZTs that contain `dummy_ask` values.
+    pub(crate) dummy_ask: Option<SpendAuthorizingKey>,
+
+    /// Proprietary fields related to the note being spent.
+    pub(crate) proprietary: BTreeMap<String, Vec<u8>>,
+}
+
+/// Information about a Sapling output within a transaction.
+///
+/// This struct is for representing Sapling outputs in a partially-created transaction. If
+/// you have a fully-created transaction, use [the regular `OutputDescription` struct].
+///
+/// [the regular `OutputDescription` struct]: crate::bundle::OutputDescription
+#[derive(Getters)]
+#[getset(get = "pub")]
+pub struct Output {
+    /// A commitment to the value created by this output.
+    pub(crate) cv: ValueCommitment,
+
+    /// A commitment to the new note being created.
+    pub(crate) cmu: ExtractedNoteCommitment,
+
+    /// The ephemeral key used to encrypt the note plaintext.
+    pub(crate) ephemeral_key: EphemeralKeyBytes,
+
+    /// The encrypted note plaintext for the output.
+    ///
+    /// Once we have memo bundles, we will be able to set memos independently of Outputs.
+    /// For now, the Constructor sets both at the same time.
+    pub(crate) enc_ciphertext: [u8; ENC_CIPHERTEXT_SIZE],
+
+    /// The encrypted output plaintext for the output.
+    pub(crate) out_ciphertext: [u8; OUT_CIPHERTEXT_SIZE],
+
+    /// The Output proof.
+    ///
+    /// This is set by the Prover.
+    pub(crate) zkproof: Option<GrothProofBytes>,
+
+    /// The address that will receive the output.
+    ///
+    /// - This is set by the Constructor.
+    /// - This is required by the Prover.
+    /// - The Signer can use `recipient` and `rseed` (if present) to verify that
+    ///   `enc_ciphertext` is correctly encrypted (and contains a note plaintext matching
+    ///   the public commitments), and to confirm the value of the memo.
+    pub(crate) recipient: Option<PaymentAddress>,
+
+    /// The value of the output.
+    ///
+    /// This may be used by Signers to verify that the value matches `cv`, and to confirm
+    /// the values and change involved in the transaction.
+    ///
+    /// This exposes the output value to all participants. For Signers who don't need this
+    /// information, or after signatures have been applied, this can be redacted.
+    pub(crate) value: Option<NoteValue>,
+
+    /// The seed randomness for the output.
+    ///
+    /// - This is set by the Constructor.
+    /// - This is required by the Prover.
+    /// - The Signer can use `recipient` and `rseed` (if present) to verify that
+    ///   `enc_ciphertext` is correctly encrypted (and contains a note plaintext matching
+    ///   the public commitments), and to confirm the value of the memo.
+    pub(crate) rseed: Option<[u8; 32]>,
+
+    /// The value commitment randomness.
+    ///
+    /// - This is set by the Constructor.
+    /// - The IO Finalizer compresses it into the bsk.
+    /// - This is required by the Prover.
+    /// - This may be used by Signers to verify that the value correctly matches `cv`.
+    ///
+    /// This opens `cv` for all participants. For Signers who don't need this information,
+    /// or after proofs / signatures have been applied, this can be redacted.
+    pub(crate) rcv: Option<ValueCommitTrapdoor>,
+
+    /// The `ock` value used to encrypt `out_ciphertext`.
+    ///
+    /// This enables Signers to verify that `out_ciphertext` is correctly encrypted.
+    ///
+    /// This may be `None` if the Constructor added the output using an OVK policy of
+    /// "None", to make the output unrecoverable from the chain by the sender.
+    pub(crate) ock: Option<OutgoingCipherKey>,
+
+    /// The ZIP 32 derivation path at which the spending key can be found for the output.
+    pub(crate) zip32_derivation: Option<Zip32Derivation>,
+
+    /// Proprietary fields related to the note being created.
+    pub(crate) proprietary: BTreeMap<String, Vec<u8>>,
+}
+
+impl fmt::Debug for Output {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        f.debug_struct("Output")
+            .field("cv", &self.cv)
+            .field("cmu", &self.cmu)
+            .field("ephemeral_key", &self.ephemeral_key)
+            .field("enc_ciphertext", &self.enc_ciphertext)
+            .field("out_ciphertext", &self.out_ciphertext)
+            .field("zkproof", &self.zkproof)
+            .field("recipient", &self.recipient)
+            .field("value", &self.value)
+            .field("rseed", &self.rseed)
+            .field("rcv", &self.rcv)
+            .field("zip32_derivation", &self.zip32_derivation)
+            .field("proprietary", &self.proprietary)
+            .finish_non_exhaustive()
+    }
+}
+
+/// The ZIP 32 derivation path at which a key can be found.
+#[derive(Debug, Getters, PartialEq, Eq)]
+#[getset(get = "pub")]
+pub struct Zip32Derivation {
+    /// The [ZIP 32 seed fingerprint](https://zips.z.cash/zip-0032#seed-fingerprints).
+    seed_fingerprint: [u8; 32],
+
+    /// The sequence of indices corresponding to the shielded HD path.
+    derivation_path: Vec<ChildIndex>,
+}